requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
email-validator==2.0.0.post2
python-dotenv==1.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
Flask-Caching==2.1.0
sentry-sdk[flask]==1.39.1
PyJWT==2.8.0
